fix(template): Move app.xaml.cs to shared "AppBase"
