+++ src/SolutionTemplate/UnoSolutionTemplate.WinUI.netcore/Base/AppHead.xaml
@@ -1,5 +1,5 @@
-﻿<Application
-    x:Class="$ext_safeprojectname$.App"
+﻿<local:App
+	x:Class="$ext_safeprojectname$.AppHead"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:$ext_safeprojectname$">
@@ -8,10 +8,14 @@
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
                 <XamlControlsResources xmlns="using:Microsoft.UI.Xaml.Controls" />
-                <!-- Add resource dictionaries here -->                
-            </ResourceDictionary.MergedDictionaries>
-            <!-- Add resources here -->
+
+				<ResourceDictionary Source="ms-appx:///$ext_safeprojectname$/AppResources.xaml" />
+				
+                <!-- Include additional resource dictionaries here, or in $ext_safeprojectname$/AppResources.xaml -->				
+			</ResourceDictionary.MergedDictionaries>
+			
+			<!-- Include additional resources here, or in $ext_safeprojectname$/AppResources.xaml -->
         </ResourceDictionary>
     </Application.Resources>
 
-</Application>
+</local:App>
