+++ src/SolutionTemplate/UnoSolutionTemplate.WinUI.netcore/Base/Base.winui.netcore.vstemplate
@@ -19,8 +19,8 @@
 	</TemplateData>
 	<TemplateContent>
 		<Project TargetFileName="$ext_safeprojectname$.Base.shproj" File="UnoQuickStart.Base.shproj" ReplaceParameters="true">
-			<ProjectItem ReplaceParameters="true" TargetFileName="App.xaml.cs">App.xaml.cs</ProjectItem>
-			<ProjectItem ReplaceParameters="true" TargetFileName="App.xaml">App.xaml</ProjectItem>
+			<ProjectItem ReplaceParameters="true" TargetFileName="AppHead.xaml.cs">AppHead.xaml.cs</ProjectItem>
+			<ProjectItem ReplaceParameters="true" TargetFileName="AppHead.xaml">AppHead.xaml</ProjectItem>
 			<ProjectItem ReplaceParameters="true" TargetFileName="base.props">base.props</ProjectItem>
 		</Project>
 	</TemplateContent>
