+++ src/SolutionTemplate/UnoSolutionTemplate.WinUI.netcore/UnoSolutionTemplate.WinUI.netcore.csproj
@@ -72,6 +72,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="Views\App.cs" />
     <None Include="Base\base.props" />
     <None Include="Base\UnoQuickStart.Base.shproj" />
     <None Include="Server\Controllers\WeatherForecastController.cs" />
@@ -121,7 +122,7 @@
     <None Include="Skia.Gtk\UnoQuickStart.Skia.Gtk.csproj" />
     <None Include="Skia.WPF\UnoQuickStart.Skia.Wpf.csproj" />
     <None Include="Wasm\Program.cs" />
-    <None Include="Base\App.xaml.cs" />
+    <None Include="Base\AppHead.xaml.cs" />
     <None Include="Views\MainPage.xaml.cs" />
     <None Include="Wasm\wwwroot\web.config" />
     <None Include="Windows\app.manifest" />
@@ -148,7 +149,7 @@
     <None Include="Views\UnoQuickStart.csproj" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="Base\App.xaml">
+    <None Include="Base\AppHead.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </None>
@@ -281,7 +282,12 @@
       <SubType>Designer</SubType>
     </None>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <None Include="Views\AppResources.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </None>
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(VSToolsPath)\VSSDK\Microsoft.VsSDK.targets" Condition="'$(VSToolsPath)' != ''" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
