+++ src/SolutionTemplate/UnoSolutionTemplate.WinUI.netcore/Views/AppResources.xaml
@@ -1,0 +1,15 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+					xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+
+	<ResourceDictionary.MergedDictionaries>
+		<!--
+			Add resource dictionaries.
+
+			For example:
+			<ResourceDictionary Source="Styles/MyButtonStyles.xaml" />
+		-->
+	</ResourceDictionary.MergedDictionaries>
+	
+	<!-- Add resources here -->
+	
+</ResourceDictionary>
